New Badge Logic added, where bagde is linked to cause.

+++ app/src/main/java/com/sharesmile/share/profile/badges/SeeAchivedBadge.java
@@ -6,5 +6,5 @@ import com.sharesmile.share.AchievedBadge;
 
 public interface SeeAchivedBadge {
 
-    public void showBadgeDetails(int id,String badgeType);
+    public void showBadgeDetails(long id,String badgeType);
 }
